fix: vercel deployments failing
